role graph further stripped and test for admin role started.
