+++ core/templates/dev/head/pages/admin/roles_tab/roles_tab_directive.html
@@ -90,8 +90,7 @@
       <md-card class="oppia-page-card oppia-long-text" style="">
         <legend class="text-center">Role Hierarchy</legend>
           <role-graph graph-data="graphData"
-                        node-fill="white"
-                        graph-data-loaded="graphDataLoaded">
+                      graph-data-loaded="graphDataLoaded">
           </role-graph> 
       </md-card>
     </div>
